Added Text Rendering
It's fairly robust meaning it has a lot of different usage options, doubt I'll use the TextManager.java render calls because I'm not a fan of that style and the cost of the loop, just using the render call off the TextSource.java is smarter imo.

This method is super expensive, I need to consider combining the vertices, creating a texture atlas, etc. For now this is usable if you don't mind the 1k fps hit (Still above 2k).