TowerDefense Example Game!
New Game! TowerDefense example game added, very minimal game but it displays UI, rendering, raycasting, etc. Just shows the general way a game can be made. However it is quite roughly coded and is really just for the sake of saying I've created something in the engine

Massive overhaul of engine due to having actually created a game in it so features that were lacking, or that I didn't like how I implemented were completely redone.